Move capability event handler to main class
