+++ src/main/java/com/alan199921/astral/Astral.java
@@ -6,6 +6,7 @@ import com.alan199921.astral.blocks.FeverweedBlock;
 import com.alan199921.astral.blocks.SnowberryBush;
 import com.alan199921.astral.capabilities.inner_realm_teleporter.IInnerRealmTeleporterCapability;
 import com.alan199921.astral.capabilities.inner_realm_teleporter.InnerRealmTeleporterCapability;
+import com.alan199921.astral.capabilities.inner_realm_teleporter.InnerRealmTeleporterProvider;
 import com.alan199921.astral.capabilities.inner_realm_teleporter.InnerRealmTeleporterStorage;
 import com.alan199921.astral.dimensions.ModDimensions;
 import com.alan199921.astral.items.EnlightenmentKey;
@@ -19,9 +20,11 @@ import com.alan199921.astral.setup.ServerProxy;
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.world.World;
 import net.minecraftforge.common.DimensionManager;
 import net.minecraftforge.common.ModDimension;
 import net.minecraftforge.common.capabilities.CapabilityManager;
+import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.DistExecutor;
@@ -55,6 +58,7 @@ public class Astral {
 
         @SubscribeEvent
         public static void init(final FMLCommonSetupEvent event){
+            System.out.println("Registering capability!");
             CapabilityManager.INSTANCE.register(IInnerRealmTeleporterCapability.class, new InnerRealmTeleporterStorage(), InnerRealmTeleporterCapability::new);
         }
 
@@ -82,6 +86,19 @@ public class Astral {
         public static void onDimensionModRegistry(final RegistryEvent.Register<ModDimension> event) {
             event.getRegistry().register(ModDimensions.innerRealm);
             DimensionManager.registerDimension(new ResourceLocation(MOD_ID, "inner_realm"), ModDimensions.innerRealm, null, true);
+        }
+
+        private static final ResourceLocation INNER_REALM_TELEPORTER = new ResourceLocation(Astral.MOD_ID, "inner_realm_teleporter");
+
+        @SubscribeEvent
+        public static void onAttachCapabilitiesToWorld(AttachCapabilitiesEvent<World> e)
+        {
+            World world = e.getObject();
+            if (world != null)
+            {
+                e.addCapability(INNER_REALM_TELEPORTER, new InnerRealmTeleporterProvider());
+                System.out.println("Attached inner realm teleporter capability!");
+            }
         }
     }
 }
